Move decorator into ContributionsEpic stories
